requirements.txt
numpy==1.22.0
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
sentence_transformers==2.7.0
skimage==0.0
streamlit==1.32.2
torch==2.3.0+cu121
